update nhan vien: upload file and review

+++ ASP.NET CORE/HRMNS/HRMNS.Application/ViewModels/HR/PhepNamViewModel.cs
@@ -15,8 +15,8 @@ namespace HRMNS.Application.ViewModels.HR
 
         [StringLength(50)]
         public string MaNhanVien { get; set; }
-        public int SoPhepNam { get; set; }
-        public int SoPhepConLai { get; set; }
+        public float SoPhepNam { get; set; }
+        public float SoPhepConLai { get; set; }
         public int Year { get; set; }
         [StringLength(50)]
         public string DateCreated { get; set; }
